requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
argilla==1.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.2
black==23.3.0
bleach==6.1.0
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama @ file:///croot/colorama_1672386526460/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
commonmark==0.9.1
contourpy==1.2.1
cryptography==41.0.1
cycler==0.12.1
dataclasses==0.6
dataclasses-json==0.5.7
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.7
docker-pycreds==0.4.0
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
faiss-cpu==1.7.4
fastapi==0.96.0
fastjsonschema==2.20.0
filelock==3.12.2
fonttools==4.52.4
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.23.5
hydra-core==1.3.2
idna==3.4
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1708996548741/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1715263367085/work
ipywidgets==8.1.3
isoduration==20.11.0
isort==5.12.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1710257277185/work
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
langchain==0.0.193
langchainplus-sdk==0.0.4
libmagic==1.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.9.0
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mistune==3.0.2
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.3
nltk==3.8.1
notebook==7.2.2
notebook_shim==0.2.4
numexpr==2.8.4
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
olefile==0.46
omegaconf==2.3.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
opencv-python==4.9.0.80
openpyxl==3.1.2
overrides==7.7.0
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pathspec==0.11.1
pbr==5.11.1
pdfminer.six==20221105
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.5.0
platformdirs==3.5.1
plotly==5.24.1
prometheus_client==0.21.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==5.27.2
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pycparser==2.21
pydantic==1.10.9
pydicom==2.4.4
pydicom-seg==0.4.1
Pygments==2.15.1
pypandoc==1.11
pyparsing==3.1.2
pyrsistent==0.20.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/python-dotenv-split_1706018097647/work
python-json-logger==2.0.7
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1705605076900/work
referencing==0.35.1
regex==2023.6.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.0.1
rpds-py==0.20.0
rt-utils==1.2.7
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.10.0
setproctitle==1.3.3
SimpleITK==2.3.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.27.0
stevedore==5.1.0
sympy==1.12.1
tabulate==0.9.0
tcia_utils==2.1.7
tenacity==8.2.2
terminado==0.18.1
threadpoolctl==3.5.0
timm==1.0.7
tinycss2==1.3.0
tomli==2.0.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4.1
tqdm==4.65.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
triton==2.3.0
typer==0.9.0
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1712329955671/work
Unidecode==1.3.8
unstructured==0.7.2
uri-template==1.3.0
urllib3==2.0.3
uvicorn==0.22.0
virtualenv==20.24.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wandb==0.17.5
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
widgetsnbextension==4.0.11
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
